requirements.txt
streamlit==1.27.0
pysqlite3-binary==0.5.2
langchain==0.0.352
langchain-community==0.0.13
langchain-google-genai==0.0.5
langchain-core==0.1.10
google-cloud-storage=2.18.0
chromadb==0.4.18
PyMuPDF==1.23.8
PyPDF2==3.0.1
google-generativeai==0.3.1